// .PRD/erp-refactor/v3/types/purchase-receipt-item.ts
export const PurchaseReceiptItemSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  barcode: z.string().optional(),
  has_item_scanned: z.boolean().optional().default(false),
  item_code: z.string(),
  product_bundle: z.string().optional(),
  supplier_part_no: z.string().optional(),
  item_name: z.string(),
  description: z.string().optional(),
  brand: z.string().optional(),
  item_group: z.string().optional(),
  image: z.string().optional(),
  image_view: z.string().optional(),
  received_qty: z.number().default(0),
  qty: z.number().optional(),
  rejected_qty: z.number().optional(),
  uom: z.string(),
  stock_uom: z.string(),
  conversion_factor: z.number(),
  retain_sample: z.boolean().optional().default(false),
  sample_quantity: z.number().int().optional(),
  received_stock_qty: z.number().optional(),
  stock_qty: z.number().optional(),
  returned_qty: z.number().optional(),
  price_list_rate: z.number().optional(),
  base_price_list_rate: z.number().optional(),
  margin_type: z.enum(['Percentage', 'Amount']).optional(),
  margin_rate_or_amount: z.number().optional(),
  rate_with_margin: z.number().optional(),
  discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  distributed_discount_amount: z.number().optional(),
  base_rate_with_margin: z.number().optional(),
  rate: z.number().optional(),
  amount: z.number().optional(),
  base_rate: z.number(),
  base_amount: z.number().optional(),
  pricing_rules: z.string().optional(),
  stock_uom_rate: z.number().optional(),
  is_free_item: z.boolean().optional().default(false),
  net_rate: z.number().optional(),
  net_amount: z.number().optional(),
  item_tax_template: z.string().optional(),
  base_net_rate: z.number().optional(),
  base_net_amount: z.number().optional(),
  valuation_rate: z.number().optional(),
  sales_incoming_rate: z.number().optional(),
  item_tax_amount: z.number().optional(),
  rm_supp_cost: z.number().optional(),
  landed_cost_voucher_amount: z.number().optional(),
  amount_difference_with_purchase_invoice: z.number().optional(),
  billed_amt: z.number().optional(),
  warehouse: z.string().optional(),
  rejected_warehouse: z.string().optional(),
  from_warehouse: z.string().optional(),
  material_request: z.string().optional(),
  purchase_order: z.string().optional(),
  purchase_invoice: z.string().optional(),
  allow_zero_valuation_rate: z.boolean().optional().default(false),
  return_qty_from_rejected_warehouse: z.boolean().optional().default(false),
  is_fixed_asset: z.boolean().optional().default(false),
  asset_location: z.string().optional(),
  asset_category: z.string().optional(),
  schedule_date: z.string().optional(),
  quality_inspection: z.string().optional(),
  material_request_item: z.string().optional(),
  purchase_order_item: z.string().optional(),
  purchase_invoice_item: z.string().optional(),
  purchase_receipt_item: z.string().optional(),
  delivery_note_item: z.string().optional(),
  putaway_rule: z.string().optional(),
  serial_and_batch_bundle: z.string().optional(),
  use_serial_batch_fields: z.boolean().optional().default(false),
  rejected_serial_and_batch_bundle: z.string().optional(),
  serial_no: z.string().optional(),
  rejected_serial_no: z.string().optional(),
  batch_no: z.string().optional(),
  include_exploded_items: z.boolean().optional().default(false),
  bom: z.string().optional(),
  weight_per_unit: z.number().optional(),
  total_weight: z.number().optional(),
  weight_uom: z.string().optional(),
  manufacturer: z.string().optional(),
  manufacturer_part_no: z.string().optional(),
  expense_account: z.string().optional(),
  item_tax_rate: z.string().optional(),
  wip_composite_asset: z.string().optional(),
  provisional_expense_account: z.string().optional(),
  project: z.string().optional(),
  cost_center: z.string().optional().default(':Company'),
  page_break: z.boolean().optional().default(false),
  sales_order: z.string().optional(),
  sales_order_item: z.string().optional(),
  subcontracting_receipt_item: z.string().optional(),
});

export type PurchaseReceiptItem = z.infer<typeof PurchaseReceiptItemSchema>;